Add files via upload
lab2_from_home

+++ lab1wp-main/src/main/java/mk/ukim/finki/wp/lab1wp/model/BookReservation.java
@@ -9,5 +9,5 @@ public class BookReservation {
     private String bookTitle;
     private String readerName;
     private String readerAddress;
-    private Long numberOfCopies;
+    private Integer numberOfCopies;
 }
